Adds provision to discourage prorietary relicensing.
Subsection 5c is modified such that this copyleft clause fails to
apply if the licensor licenses a Covered Work under Proprietary Terms
(currently undefined).

The availability of proprietary relicensing (Bradley Kuhn's term) for
licensors of a GNU GPL-licensed work with copyright ownership
consolidated in one entity is arguably a significant flaw in the GNU
GPL. It has been assumed to be unavoidable as a consequence of the GNU
GPL being a copyright license. I have spoken previously about the
reputational harm to strong copyleft caused by proprietary relicensing
business models and the social harm that results from overreaching
interpretations of the GNU GPL that such business models have encouraged.