Changed "price" to "fee"; deleted 2nd paragraph of verbatim copies section.
To my recollection, "price" was substituted for "fee" in the course of
GNU GPLv3 drafting because of a concern that some mistakenly believed
the position of the FSF to be that one could charge for "services" but
not "sell copies" (potentially without providing services). This was
so despite language in GNU GPLv2 and draft GNU GPLv3 stating
otherwise, and despite such guidance from the FSF as
http://www.gnu.org/philosophy/selling.html.

At this point, given the changes made and the degree to which the FSF
has clarified this point (and the degree to which the right to
commercialize is understood to be fundamental to FLOSS licensing), it
seems safe to revert to using "fee". (The changes made in the previous
commit call for use of one verb for both the "sale" of a copy and the
provision of a for-pay service.)