Simplified verbatim distribution section.
Previous commits allow this significant simplification of what is now
sec. 3 (corresponding to sec. 4 of GNU GPLv3). "As you receive it" and
"in any medium" seem (at least now, given previous changes) redundant.

The affirmative requirement to publish an "appropriate copyright
notice" is deleted. This makes Copyleft.next conform with typical GNU
GPL practice in the real world. If someone receives source code from
upstream, and does not modify it, they will in nearly all cases *not*
bother to publish an "appropriate copyright notice" on the source code
that is not already present (in other words, licensees do generally
preserve copyright notices, but they don't fix upstream "failures" to
place copyright notices, and for good reason. It is typically going to
be difficult for the downstream licensee to figure out what the
"appropriate copyright notice" should be for an unmodified upstream
work, and it is unreasonable to expect the licensee to do such work.
One wonders whether this requirement is a holdover from the era before
US entry into Berne.

The requirement added in GNU GPLv3 to keep intact not only basic kinds
of legal notices but also any allowed additional requirements was
deleted in this commit. I now think my deletion (made a day or so
before writing this commit message) was in error, but the solution may
be to revise the definition of "Legal Notices".