Added explicit inbound=outbound rule based on Apache License 2.0 sec 5.
This commit adds a section based closely on the ingenious first
sentence of section 5 of the Apache License 2.0, which essentially
says it is a copyright license condition that upstream contributions
are licensed by default under the license you received. It thus
codifies the customary "inbound=outbound" rule of FLOSS projects and
arguably renders unnecessary separate CLAs or the like (or at least
defeats some of the arguments asserted in their favor).

There is something awkward here that will have to be fixed
somehow. The problem essentially has to do with precisely defining who
"We" are, given the "explicit licensor" architecture of the current
draft. It may be helpful to think about how the issue is treated in
practice in Apache License contexts.