Modifications to patent license grant section.
The current scheme may be getting too complex to be desirable or
workable, but anyway it moves the definition of "Essential Patent
Claims" to section 0, generalizes it so that it can apply to upstream
licenses You receive and downstream licenses You grant. The first
paragraph of what was GNU GPLv3 sec. 11 is deleted as unnecessary, and
a provision is added making clear that granting the same sort of
patent license is required if "You" Distribute a Derivative Work. This
structure is a bit like what was used in the second discussion draft
of GNU GPLv3. But a simpler, more elegant approach would be
preferable.