Simplify structure of patent license grant.
Previous commits if anything added complexity to the architecture of
the license draft in addressing the patent license grant.

A problem noted during the course of GNU GPLv3 drafting was that
existing drafts did not bind the original licensor to grant the same
patent license that (in certain such drafts) distributors were
required to grant (or make equivalent nonassert covenants).

The ultimate solution adopted in GNU GPLv3 was to have a provision in
its section 11 that paralleled the automatic licensing provision (in
its section 10), applicable to upstream GPLv3 "contributors"
(copyright-holding licensors).

In my previous commits I made things more complicated by such
mechanisms as a definition of "Predecessor Program" and so forth.

In this commit a simpler solution is proposed: just include the patent
license grant along with the basic copyright license grant in the
current form of the "basic permissions" section. The effect ought to
be similar to that of GNU GPLv3, since those who create and Distribute
Derivative Works must license the entire Derivative Work under
copyleft-next, which means they must grant the copyright and patent
licenses set forth in the basic-permissions section.

This allows some simplification of the text.