Merge sections 4 and 5 for simplicity.
Section 4 (on verbatim source distribution) having been quite
simplified, it seems sensible to fold it into the following more
complex section on distributing Derivative Works in Source Code form.

At the same time, there is some simplification of the language. For
example, the notion that it is necessary to add an explicit
requirement concerning section 7 additional requirements seems of
dubious importance. If a licensor does not bother to include a new
allowed additional requirement, that is the licensor's choice and no
such requirement is imposed downstream. If the licensor includes a new
allowed requirement by way of inclusion of third-party
copyleft-next-compatible material, the requirement to preserve Legal
Notices takes care of the notice problem.

Another simplification is the removal of a requirement to include a
"relevant date" (GNU GPLv2: "the date of any change") when one
provides notice of a modification. This is better treated as a
normative practice as it is often unobserved in the real world. I do
not question the importance of good 'legal' coding practices, but just
because they are good doesn't mean they should be license conditions,
particularly if they will very often be ignored.