Delete UI Legal Notices requirement in source distribution section.
This requirement, a form of which has been in the GNU GPL since at
least GNU GPLv1, really makes no particular sense in a section devoted
generally to the permission to distribute source code (for many
conceivable source code derivative works this would seem to be an
unreasonable requirement if read literally). But if it is desirable to
preserve this on policy grounds it can't be limited to the object code
distribution section. Perhaps it could be bundled into the definition
of Legal Notices somehow, but I have my doubts.

This is a good example of an issue that I think should, ideally, be
discussed/debated by strong copyleft-favoring project developers and
users. Delete for now.