Make prefatory language in "Supps" consistent; clarify compatibility.
This surely needs more thought, but the idea here is that one of these
two copyleft-next "Supps", corresponding to GNU AGPLv3 and the
anti-lockdown provisions of GNU GPLv3 section 6, can only be imposed
by the original upstream licensor. This is *essentially* the way GNU
AGPLv3 works today, only GNU AGPLv3 is presented as a separate sibling
license to GNU GPLv3 (and there are special cross-compatibility
clauses).

In earlier public drafts of GNU GPLv, there was no "Affero clause" as
such; rather, a description of a category of licenses with Affero-ish
provisions was among the several kinds of compatible additional
requirements listed today in section 7. I recall that someone once
proposed that the FSF prepare canonical license texts for the allowed
additional requirements, which for the most part would not have been
sensible, but in essence this is what was done for GNU AGPLv3 --
except that GNU AGPLv3 was written as a whole new license, mostly
identical to GNU GPLv3.

Cf. also the long-forgotten GNU GPLv2 section 8, which explicitly
legitimized the ability of the "original copyright holder" to "add an
explicit geographical distribution limitation", which, if exercised,
might not be regarded as FLOSS today. (I know of only one historical
case where this was exercised, in a context that was quite
problematic, but the possibility for abuse doesn't seem relevant to
WS-Supp and AL-Supp.)

This is not the way the anti-lockdown provisions of GNU GPLv3 work
today. There is, in fact, no entirely mandatory way to achieve what I
have currently done with AL-Supp in GNU GPLv3 today. The original
licensor can place an additional permission allowing licensees to
ignore the anti-lockdown requirements, but additional permissions can
be removed by licensees. This is the inverse, as it were.

Perhaps we need to clarify that (or whether) one can, with appropriate
copyright permission, dual-license code under copyleft-next-vanilla
and copyleft-next-AL-Supp [etc.]. This raises an important policy
question which I am deferring for now. As a related matter, I have not
addressed in copyleft-next itself the issue of compatibility of
validly "Supp"-covered material with vanilla copyleft-next
material.