Replaced "We" with "Contributor"; related improvements.
"We" is potentially confusing given that we may have just one licensor.

There continues to be a drafting complexity here because I've been
restructuring the license in a way that attempts to clarify the
licensor-licensee relationship. However a given 'Program' received by
'You' may be the result of a multitude of different ancestral
'Programs' licensed under copyleft-next.

The solution here (for now) is influenced particularly by CC BY-SA
3.0. The important automatic (direct) licensing provision of the GNU
GPL has a counterpart in at least that particular CC license. Where
the GNU GPL speaks of an automatic license to the 'Program' flowing
from possibly-remote upstream copyright holders, CC BY-SA says that
the automatic licensing of recipients of 'Adaptations' is a license to
the original Work. That approach is taken in this commit in the change
to the automatic licensing clause.

A separate change made in this commit is to move the 'Contractor
Transfer' reference from a carveout of the Distribution definition to
a carveout of the "further restrictions" concept.