Clarifies inbound=outbound rule.
'Contribution' has been renamed 'Inbound Contribution' to reduce the
potential for confusion now that 'Contributor' is being (re)used to
mean something that is generally distinct, though a licensee bound by
the inbound=outbound clause could subsequently become a part of what
'Contributor' means.

Moreover, previously a 'Contribution' had to be 'modification to the
Covered Work'. An issue here is that Covered Work now ==
{Program|Derivative Work by the licensee}. The revised definition is
more general, stating that an Inbound Contribution is a 'work
... intended for inclusion in a modified or successor version'. I have
considered whether 'Derivative Work' should be a more generalized
concept (Derivative Work by upstream as well as by 'You') and have
already had, in some earlier versions, additional defined terms
attempting to get at that idea.

Separately, the inbound=outbound rule is clarified so that the
exceptional case (where the licensee-contributor indicates other
terms) must be made at the time of submission of the upstream
contribution.