Rough merge of additional terms and no-further-restrictions sections.
Given that the additional terms section no longer addresses additional
permissions, it is an elaboration on no-further-restrictions, as
indeed is liberty-or-death which was merged with
no-further-restrictions in an earlier commit. In theory it seems
sensible to try to combine them all in one section. The current
combination is quite undesirably lengthy, however.