Clarified and narrowed anti-proprietary-relicensing clause.
This change avoids the ambiguity of verb "license". In addition, the
clause is now triggered only by the Contributor "commercially"
offering the Covered Work under Proprietary Terms. A noncommercial
offering either seems nonproblematic from a policy standpoint, or
unrealistic with respect to the particular problem this clause
addresses, or both.