Delete Legal Notices definition; improvements to sec. 5.
This commit deletes the Legal Notices definition (which I now think is
broader than necessary, and is now only used in one clause
anyway). The requirement to "keep intact notices" is made a bit
narrower (though consistent with how the corresponding requirement in
GNU GPLv2/GPLv3 should be understood, I think). Previously I believe I
was thinking also of notice preservation requirements resulting from
inclusion of code under other license terms, but it ought not be
necessary to deal with that explicitly here.

There is still an argument for a broader "preserve legal notices"
requirement, but that can be revisited (perhaps addressed in the
no-further-restrictions section).

I have also modified section 5 so that it no longer is specific to
source code distribution. This is not inconsistent with the
traditional approach taken in GNU GPL (where a distributor of a binary
has permission to distribute "under the terms of" the source code
distribution sections, along with additional conditions).

I remove the requirement of including prominent notice of modification
in the case of Derivative Works. This decision can be debated, but it
reflects common practice. It may be a good example of an aspirational
or normative rule that shouldn't necessarily be stated as a copyright
condition.

Finally a few editorial changes are made to the heading and beginning
of section 6.