Initial pass at simplification of object code distribution section.
This commit makes an initial attempt to reduce the prolixity of
section 6, but with care to preserve what specific features seem
essential. Some policy changes are made. For example, the distributor
of a physical product can comply by providing notice of a network
location from which the Corresponding Source may be obtained, instead
of being limited to either use of the offer or providing Corresponding
Source along with the product. The simplified language of the option
for network distribution of binaries is influenced by my historical
research into clarifications made by RMS regarding how the
corresponding GNU GPLv2 provision mapped to use of FTP in
circumstances that were present in the early 1990s. Much of the detail
in this part of the section, not present in GNU GPLv2 but having some
basis in (I believe) pre-GNU GPLv3 versions of the GNU license FAQ, is
really not essential.

The offer option is not limited to cases where the distributor is
shipping object code in a physical product (cf. GNU GPLv2 section 3).

I have deleted entirely the "pass on a copy of the written offer to
your friend" provision. It is clear from a comparison of the
counterpart clauses in GNU GPLv2 and GNU GPLv3 that the FSF wished to
narrow the scope of this provision somewhat. I question whether it is
necessary at all (given, e.g., the likely scope of 'distribution' or
the equivalent under applicable copyright law and likely enforcement
patterns). It is not unreasonable to impose upon licensees who might
fall under the GNU GPLv3 counterpart provision the burden of either
satisfying section 6 in some other way or of taking on the risk that
the upstream licensors are unlikely to police such activity.

These changes to section 6 will require further refinement, and debate
over the policy decisions I have made here is welcomed.