Improvements to beginning of no-further-restrictions section.
Title changed to reflect that the section takes over role played by
GNU GPLv3 in articulating theory of (inbound) license compatibility.

The Apache-style patent-termination-triggering-patent-litigation
language is preserved, even though one might argue with other changes
it is not necessary, since an explicit patent termination clause has
become de rigeur. Might be better to make this more clearly tied to
the termination provision; the fact that GNU GPLv3 has Apache-like
patent termination (though broader in that copyright licenses can be
terminated as well as patent licenses) is something that has on
occasion had to be spelled out to those trying to understand the
license.

But the third paragraph takes a rather different approach to
articulating inbound license compatibility theory than GNU
GPLv3. Instead of trying to codify GNU GPLv2 interpretive tradition
with formulations of generalized categories (along with the extra
stuff, notably the addition of a category modeled on Apache License
2.0 section 9 upstream indemnification and the limited-badgeware
detail), we do something rather simpler: list a few licenses everyone
agrees are (or in the case of Apache License 2.0, should be)
GPL-compatible, and let that explicitly-illustrative list guide
interpretation. We may wish to add one or more additional licenses to
this list. It would seem that no copyleft license could be included
here without careful examination.