Delete removability-of-GPL-incompatible-restriction clause, etc.
In addition to some minor changes, mostly in the
no-further-restrictions section, this commit deletes the
cleverly-conceived clause of GNU GPLv3 section 7 which authorizes the
licensee to remove additional terms that are "further restrictions".

The rationale for this clause is made clear in historical
documents. See, e.g. GPLv3 Third Discussion Draft Rationale, section
4.2:

   We have also clarified two clauses in section 7 that concern the
   consequences of placement of a non-allowed additional requirement
   on a work.  Draft 2 introduced a clause that authorizes recipients
   to remove a non-allowed additional requirement that the work
   purports to impose. The kind of case contemplated by this clause is
   that of a program that explicitly purports to be licensed under the
   GPL along with a supplementary restriction, such as a prohibition
   on commercial use.

To my knowledge, in the real world of GNU GPLv3/GNU AGPLv3 experience,
this invitation to remove incompatible additional terms has generally
not been (explicitly) taken advantage of in situations where it ought
to have been applicable.

Perhaps a better approach to dealing with the legitimate problem
addressed by this provision is to attempt to reason with the licensor,
or to shun the licensor's software altogether.