Further improvements to no-further-restrictions section.
Here we delete the clause derived from GNU GPLv3 section 7 which
clarifies that additional terms under that section can either take the
form of an integral license text or a sort of one-off exception or
extra condition. This no longer seems worth pointing out (indeed there
may be a justification for not doing so, as it is probably good policy
not to go out of our way to encourage tacking on of nonstandard
standalone GNU-GPL-compatible conditions to GNU GPL-licensed software).