Restored separate liberty-or-death section.
The no-further-restrictions section is rather long with
liberty-or-death (including child-of-downstream-shielding) combined
with it. Moreover, as suggested in one of the earliest commits to
copyleft-next, the heading "Liberty or Death" is too good not to use.

If the no-further-restrictions section, and/or the now-separate
liberty-or-death section, are streamlined further, the question of
combining the sections should be revisited.