Some changes in liberty-or-death.
The two provisions in the copyleft-next liberty-or-death section
(corresponding to GNU GPL liberty-or-death and GNU GPLv3 section 11
paragraph 5, aka 'child-of-downstream-shielding') should be reconciled
somehow (or else perhaps split into different sections). In this
commit I have modified the beginning of the second paragraph so that
the second paragraph is not presented as a kind of exception to the
first paragraph.