Revise distribution section to allow providing URL of license text.
In practice this is often done by FLOSS projects in place of bundling
an actual copy of the license. One may object to this liberalization
by noting that a licensee without practical net access might not get
adequate notice of terms. This scenario becomes increasingly
unrealistic. The Creative Commons licenses (at least in the 3.0
generation) allow such means of providing notice of licensing terms.