Added to legal notices requirement preservation of textual author attributions.
Versions of the GNU GPL have required that 'appropriate copyright
notices' be included, but have never required preservation of author
attributions (though licenses with reasonable author attribution
requirements have been treated as compatible, and this is made
explicit in GNU GPLv3 section 7). In reality, copyright notices in
FLOSS development often serve an attribution function. It is good
policy to encourage the use of actual author attribution notices
(instead of copyright notices) if the goal is, in fact, attribution.