Substantial simplifications to termination provision
GNU GPLv3 modified the automatic termination of GNU GPLv2 with
detailed language providing for, in effect, two different kinds of
cure opportunities. The reasons for the existence of two cure clauses
are best seen as non-policy-related consequences of the complex
drafting history of GNU GPLv3. The result today is a provision that is
more complex than necessary. (It may be noted, however, that Mozilla
chose to adopt the GNU GPLv3 approach in MPL 2.0 section 5.1.)

It may be that there is some difference in nature between the 60-day
provision (which does not use the term 'cure') and the
later-introduced 30-day cure provision that supports preservation of
both.

The revised version here dispenses entirely with stating a specific
time period.

The feature of automatic termination coupled with the possibility of
license reinstatement was a late introduction to GNU GPLv3, reflecting
concerns about the impact of earlier language on ease of license
enforcement in Germany. This feature is preserved here.