Add severability clause.
No published draft of GNU GPLv3 included a general severability
clause, but the FSF noted publicly that it had considered such
inclusion in early drafts. The FSF chose not to do so because it was
thought that the prospect of a judicially-altered GPL was worse than
any benefit a severability clause might bring. (A special-purpose
severability clause was present in GNU GPLv2 liberty-or-death but was
removed in the corresponding section of GNU GPLv3.)

Severability clauses are fairly common in legal instruments and are
found in some well-known FLOSS licenses (e.g. MPL, EPL). It seems to
me that the potential benefits of inclusion of such a clause on the
legal robustness of the license outweigh the original concerns.