Delete child-of-downstream-shielding.
In the drafting history of GNU GPLv3, this provision (which is of
great interest to students of GNU GPL history) began its life as a
simple and very broadly worded requirement to "shield" downstream
recipients from third-party patent claims licensed to the distributor
and as to which the distributor distributed "knowingly relying" on the
patent license. It met with significant resistance from some
patent-holding vendors involved in the drafting process and was
substantially narrowed into the present-day provision. This provision
contains a narrow definition of "knowingly relying" and is satisfied
merely by ensuring that source code is available free of charge on the
net.

The narrowness of the provision makes its utility questionable in view
of its length. It should not be abandoned altogether, but rather, as
with the deleted provision corresponding to GNU GPLv3 section 11
penultimate paragraph, it needs to be considered in some broader
policy exploration of how copyleft-next should deal, if at all, with
the issue of third-party patent licensing of distributors. As with the
other provision, this is a "back to the drawing board" situation.