Remove obvious clarification from 'Distribute' definition.
The removed clause was adapted from the GPLv3 definition of
'convey'. This language was added during the GPLv3 drafting process to
eliminate any doubt that GPLv3 would not in itself have an 'Affero'
requirement. The clarification was justifiable for GPLv3 given the
political atmosphere at the time coupled with the exoticness of the
'convey' term. In copyleft-next, however, a more familiar copyright
law/software licensing term is used for the defined term, and arguably
the definition is more straightforward. I suppose one could raise
questions over the scope of "make available" (though the same applies
to 'distribution') but the issue is not whether the activity is more
than "mere interaction".