Deleted dl-supp.
This deletes the supplementary anti-device-lockdown provision, whichs
is largely copied from the second half of GNU GPLv3 section 6 without
change.

I am not sure at this point whether the GPLv3 anti-lockdown provision
can be improved upon. For those who care about this policy, GPLv3
continues to be available, and copyleft-next is relicensable under
GPLv3.

If, later on, it seems desirable to add back something like dl-supp,
it will be easy to do so.

For purposes of interpretation of vanilla copyleft-next, nothing
should be read into either the earlier removal of dl-supp from the
main body of the license or the decision now to delete dl-supp
altogether. (I have actually considered putting a statement to that
effect into copyleft-next, based on a comment by Alan Cox on LKML when
Luis Rodriguez posted some patches to copyleft-next there.)