Deleted 1st half of anti-anti-circumvention provision.
As public documents indicate, the first half of GPLv3 section 3 was
aimed particularly at US anticircumvention law. The basic idea here
was the hope that courts would be influenced by a general declaration
by the licensor that covered works are in some fundamental sense not
'effective technological protection measures'.

There is no known history of use of this provision by those defending
against invocation of anti-circumvention law, let alone successful
use.

The provision is not quite saying "you, licensee, may not create a
derivative work that is designed for use as a TPM" or "you, licensee,
may not use the Program as a TPM". Presumably it would be inconsistent
with the free software definition and the 'no further restrictions'
principle to impose such conditions. Assuming that is so, the effect
of this provision cannot be very strong. Given that concern, the
apparent lack of use (thus far) of the provision, and the likely
narrow set of circumstances under which it might be supposed to be
successfully used, deletion seems appropriate.

The second paragraph (which was originally drafted with EU
anticircumvention law in mind) seems more worthy of preservation,
though it may need reexamination as well. It may be useful to see
what, if anything, is done in this area by the Creative Commons 4.0
licenses. If we imagine anticircumvention law as giving rise to
something like an exclusive property right, it would make complete
sense for a copyleft license to require broad licensing of such rights
to those downstream.