Improvements to 'Corresponding Source' definition.
On the deletion of "or generally available free programs which are
used unmodified ...": This clarification is unnecessary and
potentially misleading. It should not matter whether such a program is
'free' or not. Consider a stock proprietary compiler used in
generating the Object Code. The policy behind this clarification
sentence is just as applicable to such a nonfree program as it would
be to a stock free compiler. On the other hand, one can imagine an
effort to circumvent copyleft by somehow putting improvements to the
work in question in a modified (or new) compiler. In such a case, it
still doesn't matter whether the compiler is free or nonfree (even
with a free compiler, if it is not Corresponding Source then the user
will not get source code for the improvements to the instant
work). For related reasons, the 'unmodified' clarification is not
needed. In the case where something has been specially modified so
that it is in some sense part of the instant work, it is no longer
'generally available' (consider a special internal fork of GCC
designed to circumvent the copyleft requirement) and it is
questionable whether the modified program is 'free' if its source code
has not been made public.