Handling extraordinary private communications via publication.
Expecting all participants to never have a private conversation about the
project seems too difficult to follow.  Notwithstanding the "extraordinary
cases" exception, it seems likely that impromptu conversations will
nevertheless occur.

Therefore, this additional clause handles the inevitable case of what to
do when such a private conversation occurs: namely, publish some sort of
summary or the like in a public forum.

The public forum type is left as merely "suitable", since the project may
or may not have a mailing list (copyleft-next does, but others might not,
as this rule allows for it but does not require it).

Signed-off-by: Bradley M. Kuhn <bkuhn@ebb.org>