Simplified and clarified termination provision.
In response to some discussion on the mailing list, I have replaced
"reasonable period" for cure with a 30-day period.

Separately, I have now eliminated the remaining vestiges of the
automatic termination/provisional-to-permanent automatic reinstatement
structure of the GPLv3 provision. As noted in an earlier commit, this
was introduced late in GPLv3 drafting in response to concerns by those
experienced in GPL enforcement in Germany who apparently found the
automatic termination feature to provide certain procedural advantages
(as I recall). I am open to restoring the automatic
termination/reinstatement structure if a full explanation of why it is
desirable is made available. Without such explanation, I can only see
this as more unnecessary complexity.

It is possible that this alteration to the termination provision
raises some further issue, and I expect it will need refinement.