Deleted liberty-or-death example.
Much could be written about the example in GPLv3 section 12 and the
different example it replaces in GPLv2 section 7. I will say here
that, looked upon with fresh eyes, the copyleft-next liberty-or-death
provision seems clear enough on its own, and I see no need for an
example at this time. I expect that the history of the GNU GPL,
including the transformation from GPLv2 section 7 to GPLv3 section 12,
will inform interpretation of this provision in copyleft-next.