Placeholder for EPL compatibility provision.
One of the FSF's original public goals for GPLv3 in early 2006 was
compatibility with the Eclipse Public License and the Apache License,
both of which had already come to be associated with important free
software project ecosystems.

The FSF may have originally assumed that the main obstacle to
achieving compatibility for both licenses was the scope of a new
patent license grant and any "patent retaliation clause" (as the FSF
generally called such patent termination or defensive suspension
provisions as are found in those two licenses). It turned out to be
more complicated. With the EPL, compatibility was never achieved. The
EPL is a weak copyleft license (unlike the Apache License) and
contains a number of provisions customarily considered (at least by
the FSF) to be GPL-incompatible.

Both the FSF and the Eclipse Foundation later (in 2010) reasserted the
view that GPLv2/GPLv3/AGPLv3 and the EPL are incompatible.

I would like to have a provision that solves this issue in some simple
way. I believe it would probably have to be something like GPLv3
section 13, and it would have to be somewhat at the mercy of
interpretation of the EPL's copyleft scope.