Various substantive and nonsubstantive changes.
Minor improvements to various definitions.

Removed definition of "You" and made "this License" alias of
copyleft-next in title.

General rearrangement of provisions.

Removed definition of "Licensable" as inherent in the term.

"Keep intact" requirement now refers to "legal notices" generally.

Deleted the "Inbound=Outbound" provision. I currently consider it more
difficult to get this right than is worthwhile. In many cases the rule
will be a result of the copyleft requirement, and I have argued that
inbound=outbound is an established custom in FLOSS legal culture.

Removed the ancient clarification that you can charge for distribution
or services (GPLv3: support or warranty) as obvious. It may be more
necessary in the GNU licenses for historical-baggage reasons. But I
have removed it primarily to save space.

Added parethetical in Object Code Distribution section that 'scripts
and instructions' are part of Corresponding Source "even if previously
unrecorded", to get at bkuhn's suggestion of using "know-how".

Folded liberty-or-death into no-further-restrictions again. Perhaps
more significantly, deleted the second sentence of the
liberty-or-death as an obvious consequence of the first sentence. I
then rephrased the first sentence.

Deleted, with sadness, the only remaining descendant of GPLv3 section
7 (the clause clarifying when licenses that permit "relicensing" under
the instant copyleft license are compatible provided that the
incompatible restrictions in the license do not survive). One reason
for deleting this is that the nonsurvival may be implicit on a fair
reading of the otherwise-incompatible license (i.e. it is probably too
much to expect licenses whose drafters did not intend them to be
GPL-incompatible to explicitly address the subject of restriction
survival when they have bothered to include an *explicit* relicensing
clause). Moreover, this is an academic issue to address (unless the
FSF were to adopt copyleft-next for a future version of the GNU GPL,
or world license domination were actually achieved).

Decided against attempting EPL compatibility for now (see discussion
on mailing list).

Took out the clause descended from the last part of GPLv3 section 10
(giving rise to a kind of patent peace provision) and instead adapted
the MPL 2.0 provision into a second part of the termination
provision. This is actually a significant policy change, as the MPL
2.0 provision is in most respects broader than the counterpart in
GPLv3 (though it is narrower in attempting to focus [to some degree]
only on non-defensive patent litigation).