Modify Or-Later to give authorization to FSF and OSI.
The idea behind this revision to the Or-Later section is to add a
"check and balance" that would not otherwise be present if what I now
call the "Copyleft-Next Project" has sole power to authorize new
versions of the license. With this revision, the Copyleft-Next Project
continues to have authority to draft new versions (numbered releases),
but such new versions will be effective for purposes of the "Or-Later"
provision only if *both* the FSF and the OSI give their approval to
this.

What may still be unclear is that I don't simply mean the FSF and OSI
will recognize the new version as a free software and 'Open Source'
license, respectively (though a weaker version of this provision could
be so limited). Rather, I propose that the FSF and OSI must agree that
the new version produced by the Copyleft-Next Project is acceptable as
a new version of copyleft-next.

(No, I have not yet consulted anyone at the FSF or OSI about giving
them this check-and-balance power. :)