Reduce 'minimum period' for source availability to two years.
This applies only to the circumstance in which one distributes object
code in a physical product and opts to make source code available on
the net. The original choice of three years was based on the 3-year
duration of the written offer option in the GNU GPL.

I'm assuming a reduction from 3 years to 2 years here is not
controversial.