Delete anti-anti-circumvention placeholder.
Prior to the recent commit replacing the provision with a [FIXME], the
anti-anti-circumvention provision read:

  Each Distributor of a Covered Work (i) waives all rights to forbid
  circumvention of technological measures achieved by exercising
  rights under this License with respect to a Covered Work, and (ii)
  disclaims any intention to limit operation or modification of the
  Received Work as a means of enforcing such Upstream Distributor's or
  third parties' legal rights to forbid circumvention of technological
  measures.

which is descended from the second paragraph of GPLv3 section 3.

I am now deleting this provision in its entirety because I do not see
what real-world scenario it remedies. I assume that copyleft-next is
likely to be used principally as a license for software, much like the
GPL family.

1) Waiver: Read 'Covered Work' as broadly as possible (under
copyleft-next, at least, this means either the particular work
received by the licensee or any 'Derived Work' of that Received
Work). (Note that currently copyleft-next does not define 'Derived
Work' such that it cannot be a lateral 'Derived Work' prepared by some
other licensee in some (possibly different) distribution chain
beginning with 'Us'.)

For this waiver to be meaningful, we would have to suppose that a
Distributor (or 'We') might decide to invoke anticircumvention civil
remedies against some third party exercising copyleft-next-granted
copyright licenses "with respect to" some Covered Work. This could
mean either that the underlying copyright-protected work is the (or a)
Covered Work, or perhaps some unrelated work copyrighted by the
Distributor/We in which the asserted effective TPM is a Covered
Work.

Are these scenarios *possible*? Yes. Do they seem so likely to justify
addition of a section of the license to deal with them? At the moment
my response is 'no', but perhaps there are realistic examples I am
overlooking.

2) Disclaimer of Intent: My justification of the deletion of the
language descended from GPLv3 section 3 paragraph 1 applies here as
well. It is not clear that a 'disclaimer of intent' does anything at
all. I am not aware of anti-circumvention law statutes that give
weight to the intent of a distributor of technology that is used in a
TPM.

We will restore this provision (in improved form) if people come
forward with information suggesting that the waiver and/or disclaimer
of intent are worth restoring.

But if the only purpose of these provisions is, say, a political
statement of opposition to anti-circumvention laws, or, say, the fact
that they form an important addition to GPLv3 not present in GPLv2,
that is not a good enough reason to keep them in.

I am aware that some of my logic above might be applied by some to
other provisions, such as the patent license grant. But the difference
there is that there is now more or less a consensus that FLOSS
licenses should have some sort of patent license grant, however
minimal or maximal. No FLOSS license other than the GPLv3 family has
attempted explicitly to address anticircumvention laws since they were
introduced.

Finally, it seems to me that the 'No Further Restrictions' clause and
the copyright license grant ought to provide some of the safeguards
that the anti-anti-circumvention provisions have aimed to achieve;
and, as with some other deletions of GPLv3-specific provisions, the
outbound compatibility with (A)GPLv3 means that copyleft-next code
should satisfy those who lament the deletion of these provisions from
copyleft-next.